+++ lib/shopify_sheet_platform_interface.dart
@@ -1,4 +1,5 @@
 import 'package:plugin_platform_interface/plugin_platform_interface.dart';
+import 'package:shopify_sheet/shopify_sheet.dart';
 import 'package:shopify_sheet/shopify_sheet_method_channel.dart';
 
 abstract class ShopifySheetPlatform extends PlatformInterface {
@@ -18,8 +19,14 @@ abstract class ShopifySheetPlatform extends PlatformInterface {
     throw UnimplementedError('platformVersion() has not been implemented.');
   }
 
-  Future<void> launchCheckout(String checkoutUrl) {
-    throw UnimplementedError('launchCheckout() has not been implemented.');
+  Future<void> launchCheckout(
+      String checkoutUrl, {
+        ShopifySheetConfiguration? configuration,
+      }) {
+    return ShopifySheetPlatform.instance.launchCheckout(
+      checkoutUrl,
+      configuration: configuration,
+    );
   }
 
   /// Stream of checkout events