+++ CHANGELOG.md
@@ -17,5 +17,5 @@
 * update checkout-sheet-kit-android 3.5.2
 
 ## 0.0.7
-* update checkout-sheet-kit-android 3.5.3 + dependencies
+* update checkout-sheet-kit-android 3.5.3 + dependencies Android iOS plus added checkout sheet configuration